requirements.txt
clickhouse-connect>=0.7
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
jupyter>=1.0
jupyterlab>=4.0
python-dotenv>=1.0
umap-learn>=0.5
